requirements.txt
certifi==2018.11.29
chardet==3.0.4
cx-Oracle==7.1.3
future==0.17.1
idna==2.8
numpy==1.16.2
pandas==0.24.1
PyHive==0.6.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
six==1.12.0
tamr-unify-client==0.3.0
thrift==0.11.0
urllib3==1.24.1
varyaml==0.0.3